requirements.txt
pynput
keyboard
pystray
pillow
comtypes
setuptools<81.0
